Changed AuthContext implementation structure
