+++ src/contexts/AuthContext/useLoginAttempt.ts
@@ -1,40 +1,12 @@
 import React from 'react'
-import { createSafeContext } from './utils'
 import { AuthServicesHooks } from '@/api/auth'
-import { storageRead, storageWrite, storageDelete } from '@/utils/storage.utils'
-import { MOCK_TOKEN, STORAGE_KEY_SESSION_TOKEN } from '@/config/constants'
-import noop from 'lodash/noop'
-import { useNavigateRouter, useCurrentRoute } from '@/router'
-import { useLoadingOverlay } from './LoadingOverlayContext'
+import { useCurrentRoute, useNavigateRouter } from '@/router'
 import { useTranslation } from 'react-i18next'
+import { useLoadingOverlay } from '../LoadingOverlayContext'
+import { storageRead, storageWrite } from '@/utils/storage.utils'
+import { MOCK_TOKEN, STORAGE_KEY_SESSION_TOKEN } from '@/config/constants'
 
-const { useContext, Provider } = createSafeContext<{
-  token: string | null
-  clearToken: VoidFunction
-}>('AuthContext', {
-  token: null,
-  clearToken: noop,
-})
-
-const AuthContextProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-  const [sessionToken, setSessionToken] = React.useState<null | string>(null)
-
-  useLoginAttempt(sessionToken, setSessionToken)
-
-  const clearToken = React.useCallback(() => {
-    storageDelete(STORAGE_KEY_SESSION_TOKEN)
-    setSessionToken(null)
-  }, [])
-
-  const providerValue = React.useMemo(
-    () => ({ token: sessionToken, clearToken }),
-    [sessionToken, clearToken]
-  )
-
-  return <Provider value={providerValue}>{children}</Provider>
-}
-
-function useLoginAttempt(
+export function useLoginAttempt(
   sessionToken: string | null,
   setSessionToken: React.Dispatch<React.SetStateAction<string | null>>
 ) {
@@ -97,5 +69,3 @@ function useLoginAttempt(
     asyncLoginAttempt()
   }, []) // eslint-disable-line react-hooks/exhaustive-deps
 }
-
-export { useContext as useAuthContext, AuthContextProvider }
